# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "智能文档解析服务",
    "description": "\n把 PDF / Word 文档丢进来，用大模型抽取成结构化 JSON。\n\n- **多模板**：内置合同、简历、发票三套模板，也支持自定义 JSON Schema\n- **实时进度**：通过 SSE 推送解析与抽取进度，可看到模型逐字生成的内容\n- **成本可见**：每个任务都记录 token 用量与实际花费\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/documents": {
      "post": {
        "tags": [
          "文档"
        ],
        "summary": "上传文档（1 份或多份）",
        "description": "接收 1~N 个文件，逐个校验、解析并入库。\n\n**部分成功语义**：单个文件失败不会让整批回滚，失败项会出现在 `rejected` 中。\n内容完全相同的文件不会重复存储，会出现在 `duplicate` 中并复用已有记录。\n\n状态码固定为 200：请求本身处理成功了，每个文件的结果在响应体里分别给出。用单一状态码表达「3 个成功 1 个失败」既做不到也不直观。",
        "operationId": "uploadDocuments",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadDocuments"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "文档"
        ],
        "summary": "文档列表",
        "operationId": "listDocuments",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码，从 1 开始",
              "default": 1,
              "title": "Page"
            },
            "description": "页码，从 1 开始"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页条数，最大 100",
              "default": 20,
              "title": "Page Size"
            },
            "description": "每页条数，最大 100"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_DocumentOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "tags": [
          "文档"
        ],
        "summary": "文档详情",
        "operationId": "getDocument",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "文档"
        ],
        "summary": "删除文档",
        "description": "**不可恢复的硬删除**。外键上是 ON DELETE CASCADE，该文档下的所有抽取任务与结果会一并被数据库删除。",
        "operationId": "deleteDocument",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BatchUploadResponse": {
        "properties": {
          "accepted": {
            "items": {
              "$ref": "#/components/schemas/DocumentOut"
            },
            "type": "array",
            "title": "Accepted",
            "description": "新入库的文档"
          },
          "duplicate": {
            "items": {
              "$ref": "#/components/schemas/DocumentOut"
            },
            "type": "array",
            "title": "Duplicate",
            "description": "内容已存在，直接复用已有记录"
          },
          "rejected": {
            "items": {
              "$ref": "#/components/schemas/RejectedUpload"
            },
            "type": "array",
            "title": "Rejected",
            "description": "校验未通过的文件"
          }
        },
        "type": "object",
        "title": "BatchUploadResponse",
        "description": "批量上传结果。\n\n采用「部分成功」语义：一个坏文件不会让整批回滚。\n用户拖 20 个文件进来时，因为其中一个格式不对就全军覆没是最反直觉的行为。"
      },
      "Body_uploadDocuments": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files",
            "description": "待上传的文件，支持 .pdf / .docx / .txt / .md"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_uploadDocuments"
      },
...
      "DocumentDeleteResponse": {
        "properties": {
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          },
          "cascaded_tasks": {
            "type": "integer",
            "title": "Cascaded Tasks",
            "description": "被连带删除的抽取任务数"
          }
        },
        "type": "object",
        "required": [
          "deleted",
          "cascaded_tasks"
        ],
        "title": "DocumentDeleteResponse"
      },
      "DocumentDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "用户上传时的原始文件名，仅用于展示"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256",
            "description": "内容指纹，可用于判断是否为同一份文件"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "char_count": {
            "type": "integer",
            "title": "Char Count",
            "description": "提取到的文本字符数"
          },
          "text_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Preview",
            "description": "正文前 500 字，供列表预览"
          },
          "text_status": {
            "$ref": "#/components/schemas/TextStatus"
          },
          "text_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Error",
            "description": "提取失败时的原因说明"
          },
          "task_count": {
            "type": "integer",
            "title": "Task Count",
            "description": "该文档下的抽取任务数，删除前提示用",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "text_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Content",
            "description": "提取出的完整纯文本"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "mime_type",
          "size_bytes",
          "sha256",
          "page_count",
          "char_count",
          "text_preview",
          "text_status",
          "created_at"
        ],
        "title": "DocumentDetailOut",
        "description": "文档详情。额外带上提取出的正文，便于排查抽取质量问题。\n\n复用父类的 ``from_row``：``cls`` 在这里是 ``DocumentDetailOut``，\n因此会连同 ``text_content`` 一起映射出来。"
      },
      "DocumentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "用户上传时的原始文件名，仅用于展示"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256",
            "description": "内容指纹，可用于判断是否为同一份文件"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "char_count": {
            "type": "integer",
            "title": "Char Count",
            "description": "提取到的文本字符数"
          },
          "text_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Preview",
            "description": "正文前 500 字，供列表预览"
          },
          "text_status": {
            "$ref": "#/components/schemas/TextStatus"
          },
          "text_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Error",
            "description": "提取失败时的原因说明"
          },
          "task_count": {
            "type": "integer",
            "title": "Task Count",
            "description": "该文档下的抽取任务数，删除前提示用",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "mime_type",
          "size_bytes",
          "sha256",
          "page_count",
          "char_count",
          "text_preview",
          "text_status",
          "created_at"
        ],
        "title": "DocumentOut",
        "description": "文档摘要。不含正文——列表页拉 20 份文档的全文纯属浪费带宽。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_DocumentOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "符合条件的总条数"
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "当前页码，从 1 开始"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages",
            "description": "总页数。空结果集返回 0 而不是 1，前端据此显示\"暂无数据\"。",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "Page[DocumentOut]"
      },
...
      "RejectedUpload": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "code",
          "message"
        ],
        "title": "RejectedUpload",
        "description": "批量上传中被拒绝的文件。"
      },
...
      "TextStatus": {
        "type": "string",
        "enum": [
          "extracted",
          "no_text_layer",
          "failed"
        ],
        "title": "TextStatus",
        "description": "文档文本提取结果。"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "接口密钥。仅在 AUTH_ENABLED=true 时校验。也可用 `Authorization: Bearer <key>` 传递。",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}